requirements.txt
line-bot-sdk==3.14.0
Flask~=3.0.3
gunicorn==23.0.0
PyYAML==6.0.2
pytesseract~=0.3.10
Pillow~=10.1.0
google-cloud-vision==3.8.1
google-cloud-storage==2.18.2
protobuf~=4.25.5
requests~=2.32.3
beautifulsoup4~=4.12.3
lxml==5.3.0
openai~=1.52.2
httpx==0.27.2
google-generativeai>=0.3.0
python-dotenv>=1.0.0